backend/requirements.txt
Flask
flask-cors
python-dotenv
mysql-connector-python